Crossover started and genetic switched to CRTP. Now switching to work on random generator code until its in an acceptable state (but I need to speed up)

+++ thoth/numeric/fixed_length_bitstring.cpp
@@ -68,7 +68,7 @@ FixedLengthBitString::FixedLengthBitString(unsigned long long bits, int length)
   InitBitString(bits_, bits);
 }
 FixedLengthBitString::FixedLengthBitString(std::vector<std::byte> bitstring)
-    : bits_(bitstring), length_{bitstring.size()} {}
+    : bits_(bitstring), length_{(int) bitstring.size()} {}
 
   FixedLengthBitString::FixedLengthBitString(std::vector<std::byte> bitstring, int length)
       : bits_(bitstring), length_{length} {}
